+++ utils/priority_scheduler/priority_datatypes.h
@@ -14,7 +14,7 @@ typedef enum {
 /// Metadata used to charhterize a job type and priority.
 typedef struct _job_info {
     prio_type type; ///< The type of job.
-    int deadline; ///< The deadline at which the job must be completed.
+    double deadline; ///< The deadline at which the job must be completed.
 		void* payload; ///< The actual job data.
 } job_info;
 
@@ -25,7 +25,7 @@ typedef struct _job_info {
 typedef struct queue_conf{
 	void* queue; ///< The queue object.
 	prio_type type; ///< The queue type, see `::QUEUE_BATCH`, `::QUEUE_LOSSY`, `::QUEUE_RT`.
-	void (*enqueue)(void*,int,void*); ///< The enqueue operation used for this queue.
+	void (*enqueue)(void*,double,void*); ///< The enqueue operation used for this queue.
 	void* (*dequeue)(void*); ///< The dequeue operation used for this queue.
 	int (*check_presence)(void*); ///< Check is the queue is empty.
 	int (*check_full)(void*); ///< Check if the queue is full, this function can be NULL if the queue is infinite.
